requirements.txt
langchain-core==0.3.58
langchain==0.3.25
langchain-groq==0.2.0
langchain-community==0.3.0
langchain-milvus==0.1.10
langchain-huggingface
langgraph
pydantic==2.7.4
python-dotenv==1.0.1
sentence-transformers
pyvis
jq
hf-xet
einops
pymilvus[model]==2.5.5
langchain-mistralai
langchain-google-genai